Properly handle the adaptable argument for a Sling Model class
When Sling Models instantiation was added in SLING-9320 ([commit][1]), it did not treat the `adaptable` argument the same for all types of instantiations (Sling Models vs basic Sling Adaptable). This fixes that by rearranging the instantiation attempt order, which puts the argument provided adaptable through `ModelFactory#createModel` just like the request and resource would be.

This is a subtle semantic change, and likely warrants a minor version update.

[1]: https://github.com/apache/sling-org-apache-sling-scripting-sightly/commit/bcd46027e09182911c55bb244d42debb1bd367c7